Code improvements + support 9-bit interface
Use a 9-bit interface when using a 16-bit SSD1289 or HX8532A display.
The upper 8 bits are latched in an external 74HC373 latch. This code is
currently only implemented for the atmega32u4 (Arduino Leonardo).
Changed some char variables to uint8_t so that the effect of bitwise
operations on them is properly defined. This also reduces the code size.
Added function LCD_WRITE_COM_DATA8 to further reduce code size when
using HX8532A controller.